Implement an SSL crate
This is a crate which takes arbitrary Read/Write streams and layers SSL on top
of them. The purpose of this crate is to:

* Use native platform support instead of "always OpenSSL"
* Implement best practices in terms of verification of names and such
* Be extensible in terms of how SSL is configured and such

Not much core code had to change to implement this, although one new
abstraction, `ReadyTracker`, was added to track the read/write readiness of a
stream which can be queried by wrappers above. This is needed because SSL
streams don't actually take a readable notification and then return a readable
notification, they may in fact return a writable notification! This tracking of
readiness allows implementing these in a way that's hopefully lightweight.

Windows support not implemented yet, but I hope to get to that tomorrow.