Store TaskData in the handles
We're already making an allocation, so we can stomach the Arc. The data is still
"owned" by the task itself, which is to say it's only accessible when the right
task is present. This also means that data can go out of scope once all handles
to it have gone away, yay!

At the same time we now also have no need for id fields, we can instead use the
allocation of the Task's `Arc` as the identifying component of a task.