Remove tailcall
This optimization only rarely worked anyway and was otherwise a bit of an ugly
addition to the `Future` trait. More technically, there are two reasons tailcall
is not generally applicable:

* It *only* works in chains like `.and_then(..).and_then(..)`. If, for example,
  a `map` or `map_err` were inserted in the middle then the optimization
  wouldn't appply. Additionally, it only worked with trait objects in play,
  which isn't necessarily always the case.

* It doesn't handle "back pressure". That is if a stream of events is ready very
  quickly, there are still many test cases of it blowing the stack.

In other words, the problem that `tailcall` was supposed to solve, was still
left unsolved in most cases. Removing it also additionally unlocks removing the
`'static` bound from the trait, which is another nice benefit!