Remove the Task parameter to poll
After some re-evaluation it was concluded that this parameter isn't actually
needed as part of the `poll` function. An alternative to threading through a
task is that there's simply a thread local global which indicates the "current
task". This global is then used in the same way `thread::{park, unpark}` are
used to block and notify tasks.

Additionally, this change unlocks a few further benefits:

* The core trait and all combinators are now `#![no_std]` compatible (minus
  oneshot/channel)
* Detaches `Task` from `Future` so different runtime systems can perhaps be used
  to drive a future. (none planned currently)
* It will soon be easier to remove the `'static` bounds from the trait because
  the `Task` is stored separately and the lifetimes don't need to be connected.
* The `poll` signature in general is now much simpler as it takes 0 arguments.

A downside of this change is that it is much easier to panic a thread. All I/O
or "blocking" operations require a current task to be set in a global context,
and if not set these operations will panic. This drawback, however, seems well
worth the benefits that it unlocks.