Add {Future,Stream}::await
This commit adds an `await` function to the `Future` and `Stream` traits, both
intending to bridge the asynchronous->synchronous gap for those situations
where applicable. There's liberal documentation indicating that this is not
appropriate for event loops as it will block progress, but that's not the only
use case of futures!

Both instances will "pin" the future/stream to the stack frame it originated
from, e.g. preventing it from migrating elsewhere. This is all managed through a
custom `new_notify` task.