Add an on-by-default use_std feature to `futures`
This commit adds a new Cargo feature to the `futures` crate, `use_std`, which is
enabled by default. The `futures` crate is now tagged as `#![no_std]` with extra
support like tasks, oneshot, and channels coming in through the `use_std`
feature. Most crates will likely depend on the `use_std` version, but that
shouldn't preclude those that don't want to!

Closes #60