Lift the error type out of `Poll` into a `Result`
That is:

    type Poll<T, E> = Result<Async<T>, E>

    enum Async<T> {
        Ready(T),
        NotReady,
    }

This commit tweaks the return type of the `Future::poll` method to be more
`try!`-friendly and also more amenable to `if let` for matching against
readiness.

Additionally, `Async` is used instead of `Option` to be more clear about what's
actually happening, and it's also theoretically more extensible in the future if
need be.

Finally, the `try_poll!` macro has been removed while a new `try_ready!` macro
has been added. This new macro has *different semantics* as it propagates both
`NotReady` *and* errors to the caller. It only returns the `Async::Ready`
variant.

Closes #108