Rewrite the stream::channel implementation
The previous implementation suffered a fatal flaw (#191) where to make progress
one half was required to block literally waiting for the other. The whole point
of all the lock free shenanigans was to avoid that, so it clearly didn't solve
its intended problem!

This commit completely rewrites `stream::channel` away from this pesky `Slot`
abstraction in a way that is geared towards avoiding this race we discovered.
Lots more details can be found in the implementation itself.

Closes #191