Add a BiLock<T> type for two-way ownership
This type operates similarly to a `Mutex<T>` in terms of how we can think about
it, but it only support precisely two owners, allocating the internal `Arc`
itself. This type is intended to replace man usages of `TaskRc` where ownership
of one resources is just split between two consumers.

Implementations of `Stream` and `Sink` are available on a `BiLock<T>` which
means that if a `Stream + Sink` is split with a `BiLock` then both halves
automatically will implement those traits.