Remove the sync::spsc module for now
It no longer has as much functionality as the `mpsc` module and it's not clear
that the perf is all that different to buy two separate queue implementations.
The `spsc` queues have been moved to a `futures-spsc` crate for now and if they
prove themselves in the long run we can move them back into this crate.