Restrict the Spawn::execute function
Since rust-lang/rust#36155 has been fixed, we can avoid an extra box
allocation when calling Spawn::execute with an already boxed future.

The downside of this is that it is compatibility breaking. I am now sure
if it is worth the breakage to get the performance improvement in a few
cases.