Combinators for wrapping futures in Either.
If you are returning different futures in different branches of a
conditional expression, you need to wrap them in a construct which
unifies their types.

Currently, the easiest way to do this is to called the `.boxed()`
method on both of them, but a lower overhead solution is to wrap them
in an `Either`. This commit enables that by defining a `left` and
`right` method on `Future` which wrap the future in different variants
of the `Either`.