Zero-allocation futures
Get rid of `Arc` in the `poll` methods, replacing with a
`UnparkHandle`. Implement a custom trait object that can clone itself
by limiting the maximum size of the object. Upon `park`, instead of
giving `Arc` references we clone the contents of the `UnparkHandle`,
this generalizes the behaviour of cloning the `Arc`.