[stream] add a way to execute futures serially
I've found this to be useful when creating a list of futures upfront and then
wanting them all to be executed in the order they were passed in.

In this implementation the next future starts executing only after the previous
future is done. I believe this makes the most sense semantically, and this
avoids nasty ordering issues where the futures could be accessing a shared
resource (DB etc).