Add end_after combinator on Streams
Add an end_after combinator on Streams that ends the stream after
the element for which the provided predicate returns `true`. The
stream yields all of the elements for which the predicate returns
false and then yields the element for which the predicate returned
true as the stream's final element.

This is useful, for example, to cleanly close a stream representing
the responses from a server after the server has emitted its final
response.

In some sense, `end_after` is a counterpart to `take_while` in that
`take_while` cuts the stream before the element for which the
predicate is true, while `end_after` cuts the stream after that
element.