Tweak MPSC stress test
The test ran 10k iterations, and for each iteration it spawned a new
thread. However, it did not wait for the thread to complete before
starting the next iteration. This can result in a situation where
spawning the threads can run faster than the threads closing and
overload the system.