Implement CurrentThread executor. (#639)
The `CurrentThread` executor is an event loop that allows the user to
spawn tasks that are guaranteed to run on the current thread.

This is done by extracting the scheduling logic from FuturesUnordered
and moving it to a `scheduler` module that can be used by both
`FuturesUnordered` and `CurrentThread` executor.

This is based on the Tokio reform RFC, but also includes some API
tweaks.