Add `executor::{Sleep, Wakeup}` traits
This is the first step to supporting customizing executor sleeping
strategies. Currently, it is not possible to run a Tokio reactor and a
`CurrentThread` executor on the same thread. The executor is hard coded
to use condvars for sleeping and the Tokio reactor requires calling
`epoll_wait` (or equivalent) for blocking the current thread.

The `Sleep` trait is added to abstract over this sleeping strategy.

Beyond just supporting the Tokio reactor, adding a `Sleep` trait is
useful for integrating any logic that requires hooking into the sleep
strategy (e.g. timers).