Avoid moving future when polling from scheduler
Currently, `Scheduler` has to move every future it schedules out of the
allocated node onto the stack, then move it back when it is done
polling. This can be costly with large structs. The comment claims:

> The future was extracted above (taken ownership). That way if it
> panics we're guaranteed that the future is dropped on this thread
> and doesn't accidentally get dropped on a different thread (bad).

However, in the event of a panic. It is not possible for another thread
to be able to drop the future. This is true even if the future remains
in the node.

A node is stored in an `Arc`, which means that it is not dropped until
the ref count goes to zero. The `Bomb` structure holds a handle to the
`Arc`, ensuring that the ref count does not go to zero. The other way a
future is dropped is by calling `release_node`. This function is only
called in `Bomb::drop`, which happens in the event of a panic, or in
`Scheduler::drop`, which is permitted to drop the future as it is the
type that logically owns the future value.