Remove Weak arc usage from scheduler.
`Inner` holds an Arc reference for each node and each node holds an Arc
reference to `Inner`. This creates an `Arc` cycle, which before this
patch was broken by having `Node` use a weak ref back to `Inner`.
However, using weak references requires atomic writes when notifying the
future.

This patch removes the weak ref and exchanges it for a strong ref. The
rational is that the ref cycle is already broken by the
`Scheduler::drop` implementation. The `Scheduler` type is outside of the
ref cycle and the drop implementation iterates through each node that
`Inner` holds and drops the ref. This breaks the cycle.

`Node` now has an `Option<Arc<Inner<_>>` field. It is `Option` to
support stub nodes not having to reference `Inner` as well as will
support future improvements.