Rename `CurrentThread::block_with_init` to `run`.
The function was originally named `block_with_init` as it was intended
to be used commonly as a strategy to block the current thread waiting on
a future to complete. This functionality is now provided with
`future::blocking` and `CurrentThread` is intended to be exclusively
used as a full executor.

Given this change, the emphasis on `block_` is less important and a
function name that conveys running the executor is picked instead.