Address PR feedback
This patch addresses the feedback provided in the PR. Specifically, it
does a few things:

 * Revert the removal of the Weak arc reference.
 * Switch to trait objects to hide the `W` generic.
 * Add 'static bound to `Wakeup`.

The assumptions made when removing the `Weak` arc reference were
incorrect. The incorrect assumptions are described in the commit being
reverted (85a6351e21c1a81c59c216e3beef2ba066e0c74a) and the error is
explained [here].

[here]: https://github.com/alexcrichton/futures-rs/pull/665#discussion_r154475689

The implementation details of `current_thread` are reverted back what
they were before this PR. The only remaining differences are the naming
of the `run` function and the thread-local variable used to spawn
futures into the current thread executor now uses a trait object to hide
the `W` generic. While this adds a dynamic dispatch, this cost is only
incurred on spawn.

Finally, a 'static bound is added to the `executor::Wakeup` trait.