Reduce current_thread API surface
This patch removes the concept of daemon tasks from the current thread
executor as well as the global function to cancel all spawned tasks.

If appropriate, these APIs can be brought back at a later time, but in
an effort to initially be conservative and ship a release, they are
being held back.

The main hesitation being that these APIs are related to shutting down
an executor and do not seem appropriate to be accessible from a global
context. If they remain accessible, then once the executor
implementation is pluggable, it will require that all executors
implement the same logic.

Before making this commitment, an analysis of real use cases should be
done.