add into_inner and similar methods to StreamFuture
When selecting on a StreamFuture and other futures, it is useful to be able
to recover the underlying Stream if one of the other futures resolves first.

These methods return `Option` to account for the fact that `StreamFuture`'s
implementation of `Future::poll` consumes the underlying stream during polling
in order to return it to the caller of `Future::poll` if the stream yieleded
an element.