Fix bug in AtomicWaker
The original implementation of `AtomicWaker` required calls to
`register` to be coordinated. As such, there could be no concurrent
updates to the waker.

This requirement was removed by adding some additional internal
coordination to allow concurrent calls to `register`. However, this
change left an instance of the waker handle being used outside of the
coordination logic.

This patch fixes the bug by moving the usage of the `waker` instance
inside the lock's critical section.

This patch also reduces spurious notifications by removing the waker
once it has been notified.