Optimize no-wake path of AtomicWaker::wake
This is the second attempt, previous was unsuccessful, see #1335.

This version does a CAS with release flag in `take()` to guarantee
that change of state in `register()` is visible in waker when there's
nobody to wake.

Benchmark:

```
test unbounded_uncontended ... bench:     111,477 ns/iter (+/- 10,194) # before
test unbounded_uncontended ... bench:     104,793 ns/iter (+/- 8,565)  # with patch
```

The result is still better because when there's nobody to wake, `wake()`
does only one atomic op instead of two.

Description:

Introduce `HOLDS` bit which is true when there's waker in the cell
(option is Some).

`wake` operation immediately exits if `HOLDS` bit is not set (also
if `WAKING` bit is set which means another thread does wake right
now).

The patch includes a graphviz diagram to better understand the state
transitions.

From the diagram, it's clear that there are only 6 states although
there are 8 possible bit set values:

* EMPTY
* HOLDS
* WAKING
* WAKING|HOLDS
* REGISTERING
* REGISTERING|WAKING

Register state transitions are red, and wake state transitions are
blue.

Solid line transitions at the beginning of function calls, and
dashed lines are second transitions.