Add Sink01As03 compat shim
This adds a compat shim to go from a 0.1 Sink to a 0.3 Sink. This follows,
roughly how the stream compat works with the exception that it stores an internal
buffer of one item. This is due to the fact that there is no `LocalWaker` passed
to `Sink01::start_send`, so we can only actually write the item in the sink on the
next `Sink03::poll_read` which gets a `LocalWaker`.