Add a new futures-intrusive crate.
This crate contains a set of future-based primitives which are built on top of intrusive data structures.
The main goal of this crate is to provide types which require no dynamic allocation during runtime, independent of the number of tasks and concurrent paths that are interacting with the future.
The goal should be be to provide all types in a no-std compatible form.

Currently the crate contains a Mutex and a ManualResetEvent type, in both Sync and !Sync fashion. The !Sync variants are already no-std compatible.
Other planned types would include:
- Timers
- Channels
- More primitives (e.g. Semaphores)