Improve support for mutable borrows in select! blocks
As reported in #1811 mutable accesses were not allowed inside select!
blocks if a Future inside the same expression borrowed it. This is not
necessarily required, since the actual polling of the Futures occurs
before the result blocks are evaluated.

This change allows borrowing a variable inside a Future and referring to
the same variable inside other select branches. The change introduces
additional scoping into the select macro which separates the future poll
phase from executing the result blocks phase.

Fixes #1811