Add StreamExt::buffer_unordered_adaptable
It might be desirable to have adaptative parallelism for the
processing of buffered futures.

An example use case would be to cope with the ever changing
network conditions when sending requests to a remote host.
One might collect metrics about the performance of the
data transfer and adapt the parallelism parameters as
necessary to maximize throughput.