Optimize poll_loop! using NonNullU32
Use NonNullU32 for the limits on the number of iterations to give
the optimizer information that the first iteration will always be taken,
allowing to unroll it or move the loop exit check to the end.

To make poll_loop! macro work with a NonNullU32 argument,
apply an Into<u32> conversion to it. The iteration range is now
strictly in u32, while previously the integer type was inferred from
the macro's argument.