Rework the loop guard protocol
iteration::Policy gets methods to observe how the loop has ended.
LoopGuard, a RAII helper, is used by poll_loop! to catch early returns.
The uses of poll_loop! in futures-util needed to be changed to split
borrows out of Pin<&mut Self>, allowing the guard to borrow the
iteration checker while other state objects are used and modified
by the loop body.
Notably, this makes the code of the loops more readable.