Rotate the check in poll_loop!, remove ToYield
Perform the yield check at the end of an iteration. Document
the check-skipping effect of `continue` and the new semantics of
`break`: this can now be used to return arbitrary values from
the poll_loop! expression. This is less semantically weird than
`break ToYield`, and `continue` gets back its usual meaning of
continuing iteration unconditionally.