Simplify iteration::Policy and erase LoopGuard
It's not a good idea to let impls run arbitrary code in the RAII
destructor which may be invoked in an unwind.
Also, separation between yield_check and when_yielded is not
really necessary. Erase when_* methods on Policy and allow the
yield_check method to modify the receiver so that adaptive and
instrumentation impls are still supported.

With this, LoopGuard is not needed in poll_loop! and is gone.