Guard `run_executor` against use of park/unpark in user-code.
If user code that is run as a result of polling the futures
in the `run_executor` loop of `local_pool.rs` makes use of
the park / unpark APIs, execution may stall due to wakeups
getting "lost". This is prevented (and thereby unnecessary
calls to park / unpark avoided) through an additional
`AtomicBool` that is in full control of the code in
`local_pool.rs`.