Fix unwind safety in RemoteHandle
After analysis in #2077 I conclude that currently in remote_handle we:
- assert a user supplied future to be unwind safe and when it panics,
  we don't tear down the thread on which the panic happens.
- tear down the thread on which the handle is awaited even though no
  panic ever happened on that thread.

This behavior introduces unwind safety issues on the thread where the
panic happens and is very unusual behavior to tear down an unrelated
thread, completely different from how std::thread behaves.

This commit tries to fix that by tearing down the thread on which the
panic happens and not the thread on which the handle is awaited.

As a consequence, RemoteHandle now needs to resolve to a Result instead
of panicking. This is a BREAKING CHANGE and version bumps aren't included
here.

One issue remains to be resolved. I don't manage to make the console
print something sensible for the panicking thread.

I have removed the "run with `RUST_BACKTRACE=1`..." from the examples below because
it's not really relevant.

When trying (in the poll for Remote):
```rust
Err(e) => {
    let msg = format!("Remote future panicked with: {:?}", &e);
    drop(self.as_mut().tx().take().unwrap().send(Err(e)));
    std::panic::resume_unwind(Box::new(msg));
}
```
gives:

`thread '<unnamed>' panicked at 'Box<Any>', futures-util/examples/remote_handle.rs:11:26`

Not very helpful.

This will make it look like the thread panicked twice and still doesn't
print an actual error:

```rust
Err(e) => {
    let msg = format!("Remote future panicked with: {:?}", &e);
    drop(self.as_mut().tx().take().unwrap().send(Err(e)));
    panic!(msg);
}
```
```
thread '<unnamed>' panicked at 'Box<Any>', futures-util/examples/remote_handle.rs:11:26
thread '<unnamed>' panicked at 'Remote future panicked with: Any', <::std::macros::panic macros>:2:4
```

If we were to turn things around, sending a string to the RemoteHandle but keeping the error
from `catch_unwind` to pass it to `resume_unwind`, following the example for `resume_unwind` from
the std docs, we get:

```rust
Err(e) => {
    drop(self.as_mut().tx().take().unwrap().send(Err(Box::new("Remote future panicked"))));
    std::panic::resume_unwind(e);
}
```

`thread '<unnamed>' panicked at 'Box<Any>', futures-util/examples/remote_handle.rs:11:26`

which is no better.

# Alternatives

Currently there is no way to get the console to behave and it's actuall quite
unwieldy to do anything useful with the returned error (eg. it's not possible
to downcast it if the panic was the result of a call to unwrap or expect).

I still think that not catching the unwind at all and just returning a simple
error saying "Remote future panicked" is a viable alternative.